+++ frontend2/src/views/Queue.tsx
@@ -4,8 +4,8 @@ import type { PaginatedMatchList } from "../utils/types";
 import { useSearchParams } from "react-router-dom";
 import { getAllMatches, getScrimmagesByTeam } from "../utils/api/compete";
 import { useEpisodeId } from "../contexts/EpisodeContext";
-import BattlecodeTable from "../components/BattlecodeTable";
-import BattlecodeTableBottomElement from "../components/BattlecodeTableBottomElement";
+import Table from "../components/Table";
+import TableBottom from "../components/TableBottom";
 import Button from "../components/elements/Button";
 import RatingDelta from "../components/compete/RatingDelta";
 import MatchScore from "../components/compete/MatchScore";
